+++ src/js/const/operation-error-codes.js
@@ -1,3 +1,4 @@
 export const OPERATION_ERROR_CODES = Object.freeze({
   opCrossSelf: 'op_cross_self',
+  opAccountBlocked: 'op_account_blocked',
 })