+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+### Added
+- User account blocked error
+- Transaction error
+
 ### Removed
 - Outdated license file